// src/lib/api/types/adminAccountDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */
import type { RoleDTO } from './roleDTO';

export interface AdminAccountDTO {
  /** Creation timestamp */
  createdAt: string;
  /** Account email */
  email: string;
  /** User first name */
  firstName: string;
  /** Account identifier */
  id: string;
  /** Last login timestamp */
  lastLoginAt?: string;
  /** User last name */
  lastName: string;
  /**
   * Assigned roles
   * @nullable
   */
  roles: RoleDTO[] | null;
  /** Account status */
  status: string;
  /** Account username */
  username?: string;
}