// src/lib/api/types/roleDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface RoleDTO {
  /** Role code */
  code: string;
  /** Role description */
  description?: string;
  /** Role identifier */
  id: string;
  /** Mutable role */
  isMutable: boolean;
  /** System role */
  isSystem: boolean;
  /** Role name */
  name: string;
  /** Role type */
  type: string;
}